+++ test/test_helper.exs
@@ -1,4 +1,1 @@
 ExUnit.start()
-
-# Load all test support files
-Code.require_file("support/test_client.ex", __DIR__)